+++ src/Model/Export/Catalog/Entity/ProductVariation.php
@@ -6,6 +6,8 @@ namespace Omikron\Factfinder\Model\Export\Catalog\Entity;
 
 use Magento\Catalog\Model\Product;
 use Omikron\Factfinder\Api\Export\ExportEntityInterface;
+use Omikron\Factfinder\Api\Export\FieldInterface;
+use Omikron\Factfinder\Model\Export\Catalog\FieldProvider;
 use Omikron\Factfinder\Model\Formatter\NumberFormatter;
 
 class ProductVariation implements ExportEntityInterface
@@ -20,18 +22,23 @@ class ProductVariation implements ExportEntityInterface
     private $numberFormatter;
 
     /** @var array */
-    private $data;
+    private $configurableData;
+
+    /** @var FieldProvider */
+    private $fieldprovider;
 
     public function __construct(
         Product $product,
         Product $configurable,
         NumberFormatter $numberFormatter,
+        FieldProvider $variantFieldProvider,
         array $data = []
     ) {
-        $this->product         = $product;
-        $this->configurable    = $configurable;
-        $this->numberFormatter = $numberFormatter;
-        $this->data            = $data;
+        $this->product          = $product;
+        $this->configurable     = $configurable;
+        $this->numberFormatter  = $numberFormatter;
+        $this->configurableData = $data;
+        $this->fieldprovider    = $variantFieldProvider;
     }
 
     public function getId(): int
@@ -41,13 +48,19 @@ class ProductVariation implements ExportEntityInterface
 
     public function toArray(): array
     {
-        return array_merge($this->data, [
-            'ProductNumber' => (string) $this->product->getSku(),
-            'Price'         => $this->numberFormatter->format((float) $this->product->getFinalPrice()),
-            'Availability'  => (int) $this->product->isAvailable(),
-            'HasVariants'   => 0,
-            'MagentoId'     => $this->getId(),
-        ]);
+        $baseData = [
+                'ProductNumber' => (string) $this->product->getSku(),
+                'Price'         => $this->numberFormatter->format((float) $this->product->getFinalPrice()),
+                'Availability'  => (int) $this->product->isAvailable(),
+                'HasVariants'   => 0,
+                'MagentoId'     => $this->getId(),
+            ] + $this->configurableData;
+
+        return array_reduce($this->fieldprovider->getVariantFields(),
+            function (array $result, FieldInterface $field): array {
+                return [$field->getName() => $field->getValue($this->product)] + $result;
+            }, $baseData
+        );
     }
 
     public function getProduct(): Product