+++ src/Model/Export/Catalog/FieldProvider.php
@@ -18,20 +18,51 @@ class FieldProvider implements FieldProviderInterface
     private $fieldFactory;
 
     /** @var FieldInterface[] */
-    private $productFields;
+    private $productFieldProviders;
 
-    public function __construct(ExportConfig $config, GenericFieldFactory $fieldFactory, array $productFields)
+    /** @var FieldInterface[] */
+    private $variantFieldProviders;
+
+    /** @var array */
+    private $cachedVariantFields;
+
+    /** @var array */
+    private $cachedFields;
+
+    public function __construct
+    (
+        ExportConfig $config,
+        GenericFieldFactory $fieldFactory,
+        array $productFields = [],
+        array $variantFields = []
+    ) {
+        $this->config                = $config;
+        $this->productFieldProviders = $productFields;
+        $this->variantFieldProviders = $variantFields;
+        $this->fieldFactory          = $fieldFactory;
+    }
+
+    public function getVariantFields(): array
     {
-        $this->config        = $config;
-        $this->productFields = $productFields;
-        $this->fieldFactory  = $fieldFactory;
+        if (!$this->cachedVariantFields) {
+            $this->cachedVariantFields = $this->getFrom($this->variantFieldProviders);
+        }
+        return $this->cachedVariantFields;
     }
 
     public function getFields(): array
     {
+        if (!$this->cachedFields) {
+            $this->cachedFields = $this->getFrom($this->productFieldProviders);
+        }
+        return $this->cachedFields;
+    }
+
+    public function getFrom(array $fields): array
+    {
         return array_reduce($this->config->getSingleFields(), function (array $fields, string $attributeCode): array {
             $attribute = $this->fieldFactory->create(['attributeCode' => $attributeCode]);
             return [$attribute->getName() => $attribute] + $fields;
-        }, $this->productFields);
+        }, $fields);
     }
 }