+++ src/Model/Export/FeedFactory.php
@@ -30,9 +30,12 @@ class FeedFactory
             throw new InvalidArgumentException(sprintf('There is no feed configuration for the given type: %s', $type));
         }
 
-        $fields = is_array($this->feedPool[$type]['fieldProvider'])
-            ? $this->feedPool[$type]['fieldProvider']
-            : $this->objectManager->create($this->feedPool[$type]['fieldProvider'])->getFields();
+        $fieldProvider = $this->feedPool[$type]['fieldProvider'];
+        $fields = is_array($fieldProvider)
+            ? $fieldProvider
+            : call_user_func(function(FieldProviderInterface $fieldProvider) {
+                return $fieldProvider->getFields() + $fieldProvider->getVariantFields();
+            }, $this->objectManager->create($fieldProvider));
 
         $dataProvider = $this->objectManager->create($this->feedPool[$type]['dataProvider'], ['fields' => $fields]);